moved the mapping functionality of PatientProfileResource to PatientProfileMapper
