+++ api/pom.xml
@@ -54,6 +54,16 @@
             <type>pom</type>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.openmrs.module</groupId>
+            <artifactId>webservices.rest-omod-common</artifactId>
+            <version>2.14-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.openmrs.module</groupId>
+            <artifactId>webservices.rest-omod-1.8</artifactId>
+            <version>2.14-SNAPSHOT</version>
+        </dependency>
 
         <!-- End OpenMRS core -->
     </dependencies>
